Convert classify module to package
Split into:
- classify/tools.py: CANONICAL_TOOLS, TOOL_CATEGORY_MAP, normalize/classify
- classify/shell.py: tree-sitter AST analysis, shell command classification
- classify/__init__.py: public re-exports

No behavior changes — same public API, all 155 tests pass.
